+++ tests/testthat/test-bounds.r
@@ -12,11 +12,12 @@ test_that("rescale_mid returns correct results", {
 })
 
 
-test_that("resacle_max returns correct results", {
+test_that("rescale_max returns correct results", {
   expect_equal(rescale_max(0), NaN)
   expect_equal(rescale_max(1), 1)
   expect_equal(rescale_max(.3), 1)
   expect_equal(rescale_max(c(4, 5)), c(0.8, 1.0))
+  expect_equal(rescale_max(c(-3, 0, -1, 2)), c(-1.5, 0, -0.5, 1))
   expect_equal(rescale_max(c(-3, 0, -1, 2)), c(-1.5, 0, -0.5, 1))
 })
 
@@ -46,4 +47,11 @@ test_that("scaling is possible with dates and times", {
   expect_equal(rescale(dates, from = c(dates[1], dates[4])), seq(0, 1, 1/3))
   expect_equal(rescale_mid(dates, mid = dates[3])[3], 0.5)
 
+})
+
+test_that("scaling is possible with integer64 data", {
+  x <- bit64::as.integer64(2^60) + c(0:3)
+  expect_equal(
+    rescale_mid(x, mid = bit64::as.integer64(2^60) + 1),
+    c(0.25, 0.5, 0.75, 1))
 })