+++ NEWS.md
@@ -8,6 +8,9 @@
 
 * Switched from reference classes to R6 (#96).
 
+* `rescale()` and `rescale_mid()` are now S3 generics, and work with `numeric`,
+  `Date`, `POSIXct`, `POSIXlt` and `bit64::integer64` objects (@zeehio, #74).
+
 # scales 0.4.1
 
 * `extended_breaks()` no longer fails on pathological inputs
@@ -16,9 +19,6 @@
 
 * `train_discrete()` gets a new `na.rm` argument which controls whether
   `NA`s are preserved or dropped.
-
-* `rescale()` and `rescale_mid()` are now S3 generics, and work with `numeric`,
-  `Date`, `POSIXct`, `POSIXlt` and `bit64::integer64` objects (@zeehio, #74).
 
 # scales 0.4.0